+++ mandiant-threat-intel/operations.py
@@ -1,8 +1,9 @@
-""" Copyright start
-  Copyright (C) 2008 - 2022 Fortinet Inc.
-  All rights reserved.
-  FORTINET CONFIDENTIAL & FORTINET PROPRIETARY SOURCE CODE
-  Copyright end """
+"""
+Copyright start
+MIT License
+Copyright (c) 2024 Fortinet Inc
+Copyright end
+"""
 
 from .mandiant_api_auth import *
 from connectors.core.connector import get_logger, ConnectorError
@@ -28,7 +29,7 @@ def make_rest_call(endpoint, method, connector_info, config, data=None, params=N
         logger.debug("Token: {0}".format(token))
         logger.debug("Endpoint URL: {0}".format(url))
         headers = {'Content-Type': 'application/json',
-                   'Accept': 'application/vnd.oasis.stix+json; version=2.1',
+                   'Accept': 'application/json',
                    'X-App-Name': 'fortisoar.fortinet.v1.0',
                    'Authorization': token}
         logger.debug("Headers: {0}".format(headers))
@@ -39,6 +40,8 @@ def make_rest_call(endpoint, method, connector_info, config, data=None, params=N
             return response.json()
         elif response.status_code == 204:
             return dict()
+        elif response.status_code == 404:
+            return {"Message": "Not Found", "http_status": "404"}
         else:
             raise ConnectorError("{0}".format(errors.get(response.status_code)))
     except requests.exceptions.SSLError:
@@ -136,6 +139,25 @@ def fetch_indicators(config, params, connector_info):
         raise ConnectorError("{0}".format(str(err)))
 
 
+def get_reputation_of_indicators(config, params, connector_info):
+    try:
+        indicator_value = params.get('indicatorValue')
+        tempList = []
+        tempList.append(indicator_value)
+        endpoint = "/v4/indicator"
+        data = {'requests': [{'values': tempList}]}
+        jsonData = json.dumps(data)
+        logger.debug("JSONDATA: {0}".format(jsonData))
+        response = make_rest_call(endpoint, 'POST', connector_info, config, data=jsonData)
+        logger.debug("Response: {0}".format(response))
+        if bool(response):
+            return response
+        return {"objects": []}
+    except Exception as err:
+        logger.exception("{0}".format(str(err)))
+        raise ConnectorError("{0}".format(str(err)))
+
+
 def get_reports(config, params, connector_info):
     try:
         endpoint = "/collections/reports/objects"
@@ -215,6 +237,21 @@ def search_collections(config, params, connector_info):
         raise ConnectorError("{0}".format(str(err)))
 
 
+def execute_an_api_call(config, params, connector_info):
+    try:
+        endpoint = params.get("endpoint")
+        http_method = params.get("method")
+        query_params = params.get("query_params") if params.get("query_params") else None
+        payload = json.dumps(params.get("payload")) if params.get("payload") else None
+        logger.debug("Payload: {0}".format(payload))
+        response = make_rest_call(endpoint, http_method, connector_info, config, params=query_params, data=payload)
+        logger.debug("Response: {0}".format(response))
+        return response
+    except Exception as err:
+        logger.exception("{0}".format(str(err)))
+        raise ConnectorError("{0}".format(str(err)))
+
+
 def _check_health(config, connector_info):
     try:
         return check(config, connector_info)
@@ -228,5 +265,7 @@ operations = {
     'fetch_indicators': fetch_indicators,
     'get_reports': get_reports,
     'get_alerts': get_alerts,
-    'search_collections': search_collections
+    'search_collections': search_collections,
+    'get_reputation_of_indicators': get_reputation_of_indicators,
+    'execute_an_api_call': execute_an_api_call
 }