+++ mandiant-threat-intel/release_notes.md
@@ -1,4 +1,8 @@
-#### Following enhancements have been made to the Mandiant Threat Intelligence Connector in version 1.1.0:
+#### What's Improved
 
-- This connector version is now certified.
-- The connector now supports data ingestion using the Data Ingestion Wizard.
+Following enhancements have been made to the Mandiant Threat Intelligence Connector in version 1.2.0: 
+
+- Added following new actions:
+  - Get Indicator Reputation
+  - Execute an API Request
+